+++ CHANGELOG.md
@@ -1,2 +1,3 @@
 # 0.1.1
 - typo fix: `MCaptcha::decrement_visiotr()` became `MCaptcha::decrement_visitor()`
+- `serde::{Serialize, Deserialize}` impls (shouldn't break anything)